+++ library/LibUtilities/Foundations/NodalPrismElec.cpp
@@ -34,10 +34,7 @@
 
 #include <boost/core/ignore_unused.hpp>
 
-#include <LibUtilities/BasicUtils/ErrorUtil.hpp>
-#include <LibUtilities/BasicUtils/SharedArray.hpp>
 #include <LibUtilities/Foundations/NodalPrismElec.h>
-#include <LibUtilities/Foundations/Points.h>
 #include <vector>
 
 namespace Nektar
@@ -50,62 +47,62 @@ bool NodalPrismElec::initPointsManager[] = {PointsManager().RegisterCreator(
 
 namespace
 {
-bool isVertex(int t, int y, int npts)
+bool isVertex(size_t t, size_t y, size_t npts)
 {
     return (t == 0 && y == 0) || (t == 1 && y == 0) || (t == 2 && y == 0) ||
            (t == 0 && y == (npts - 1)) || (t == 1 && y == (npts - 1)) ||
            (t == 2 && y == (npts - 1));
 }
 
-bool isEdge_01(int t, int y, int npts)
+bool isEdge_01(size_t t, size_t y, size_t npts)
 {
     return y == 0 && t > 2 && t <= npts;
 }
 
-bool isEdge_12(int t, int y, int npts)
+bool isEdge_12(size_t t, size_t y, size_t npts)
 {
     boost::ignore_unused(y, npts);
     return t == 1;
 }
 
-bool isEdge_23(int t, int y, int npts)
+bool isEdge_23(size_t t, size_t y, size_t npts)
 {
     return y == (npts - 1) && t > 2 && t <= npts;
 }
 
-bool isEdge_30(int t, int y, int npts)
+bool isEdge_30(size_t t, size_t y, size_t npts)
 {
     boost::ignore_unused(y, npts);
     return t == 0;
 }
 
-bool isEdge_04(int t, int y, int npts)
+bool isEdge_04(size_t t, size_t y, size_t npts)
 {
     return y == 0 && t >= 3 + 2 * (npts - 2) && t < 3 + 3 * (npts - 2);
 }
 
-bool isEdge_14(int t, int y, int npts)
+bool isEdge_14(size_t t, size_t y, size_t npts)
 {
     return y == 0 && t >= 3 + (npts - 2) && t < 3 + 2 * (npts - 2);
 }
 
-bool isEdge_25(int t, int y, int npts)
+bool isEdge_25(size_t t, size_t y, size_t npts)
 {
     return y == npts - 1 && t >= 3 + (npts - 2) && t < 3 + 2 * (npts - 2);
 }
 
-bool isEdge_35(int t, int y, int npts)
+bool isEdge_35(size_t t, size_t y, size_t npts)
 {
     return y == npts - 1 && t >= 3 + 2 * (npts - 2) && t < 3 + 3 * (npts - 2);
 }
 
-bool isEdge_45(int t, int y, int npts)
+bool isEdge_45(size_t t, size_t y, size_t npts)
 {
     boost::ignore_unused(y, npts);
     return t == 2;
 }
 
-bool isEdge(int t, int y, int npts)
+bool isEdge(size_t t, size_t y, size_t npts)
 {
     return isEdge_01(t, y, npts) || isEdge_12(t, y, npts) ||
            isEdge_23(t, y, npts) || isEdge_30(t, y, npts) ||
@@ -114,37 +111,37 @@ bool isEdge(int t, int y, int npts)
            isEdge_45(t, y, npts);
 }
 
-bool isFace_0123(int t, int y, int npts)
+bool isFace_0123(size_t t, size_t y, size_t npts)
 {
     boost::ignore_unused(y);
     return t < 3 + (npts - 2);
 }
 
-bool isFace_014(int t, int y, int npts)
+bool isFace_014(size_t t, size_t y, size_t npts)
 {
     boost::ignore_unused(t, npts);
     return y == 0;
 }
 
-bool isFace_1254(int t, int y, int npts)
+bool isFace_1254(size_t t, size_t y, size_t npts)
 {
     boost::ignore_unused(y);
     return t < 3 + 2 * (npts - 2) && t >= 3 + (npts - 2);
 }
 
-bool isFace_325(int t, int y, int npts)
+bool isFace_325(size_t t, size_t y, size_t npts)
 {
     boost::ignore_unused(t);
     return y == (npts - 1);
 }
 
-bool isFace_0354(int t, int y, int npts)
+bool isFace_0354(size_t t, size_t y, size_t npts)
 {
     boost::ignore_unused(y);
     return t < 3 + 3 * (npts - 2) && t >= 3 + 2 * (npts - 2);
 }
 
-bool isFace(int t, int y, int npts)
+bool isFace(size_t t, size_t y, size_t npts)
 {
     return isFace_0123(t, y, npts) || isFace_014(t, y, npts) ||
            isFace_1254(t, y, npts) || isFace_325(t, y, npts) ||
@@ -159,7 +156,7 @@ void NodalPrismElec::v_CalculatePoints()
     PointsBaseType::v_CalculatePoints();
 
     // Populate m_points
-    unsigned int npts = GetNumPoints();
+    size_t npts = GetNumPoints();
 
     LibUtilities::PointsKey pkey1(npts, LibUtilities::eNodalTriElec);
     Array<OneD, NekDouble> u1, v1;
@@ -168,7 +165,7 @@ void NodalPrismElec::v_CalculatePoints()
     Array<OneD, NekDouble> u;
     LibUtilities::PointsManager()[pkey2]->GetPoints(u);
 
-    for (unsigned int y = 0, index = 0; y < npts; y++)
+    for (size_t y = 0, index = 0; y < npts; y++)
     {
         for (size_t t = 0; t < u1.size(); t++, index++)
         {
@@ -185,7 +182,7 @@ void NodalPrismElec::v_CalculatePoints()
 
 void NodalPrismElec::NodalPointReorder3d()
 {
-    unsigned int npts = GetNumPoints();
+    size_t npts = GetNumPoints();
     using std::vector;
     vector<int> vertex;
     vector<int> iEdge_01;             // interior edge 0
@@ -206,9 +203,9 @@ void NodalPrismElec::NodalPointReorder3d()
     vector<int> map;
 
     // Build the lattice prism left to right - bottom to top
-    for (unsigned int y = 0, index = 0; y < npts; y++)
+    for (size_t y = 0, index = 0; y < npts; y++)
     {
-        for (unsigned int t = 0; t < npts * (npts + 1) / 2; t++, index++)
+        for (size_t t = 0; t < npts * (npts + 1) / 2; t++, index++)
         {
             if (isVertex(t, y, npts))
             {
@@ -292,104 +289,104 @@ void NodalPrismElec::NodalPointReorder3d()
     std::reverse(iEdge_35.begin(), iEdge_35.end());
 
     // faces
-    for (unsigned int i = 0; i < npts - 2; i++)
+    for (size_t i = 0; i < npts - 2; i++)
     {
-        for (unsigned int j = i + 1; j < npts - 2; j++)
+        for (size_t j = i + 1; j < npts - 2; j++)
         {
             std::swap(iFace_1254[i * (npts - 2) + j],
                       iFace_1254[j * (npts - 2) + i]);
         }
     }
-    for (int i = 0; i < npts - 2; i++)
+    for (size_t i = 0; i < npts - 2; i++)
     {
         std::reverse(iFace_0354.begin() + (i * (npts - 2)),
                      iFace_0354.begin() + (i * (npts - 2) + npts - 2));
     }
-    for (unsigned int i = 0; i < npts - 2; i++)
+    for (size_t i = 0; i < npts - 2; i++)
     {
-        for (unsigned int j = i + 1; j < npts - 2; j++)
+        for (size_t j = i + 1; j < npts - 2; j++)
         {
             std::swap(iFace_0354[i * (npts - 2) + j],
                       iFace_0354[j * (npts - 2) + i]);
         }
     }
 
-    for (unsigned int n = 0; n < vertex.size(); ++n)
+    for (size_t n = 0; n < vertex.size(); ++n)
     {
         map.push_back(vertex[n]);
     }
 
-    for (unsigned int n = 0; n < iEdge_01.size(); ++n)
+    for (size_t n = 0; n < iEdge_01.size(); ++n)
     {
         map.push_back(iEdge_01[n]);
     }
 
-    for (unsigned int n = 0; n < iEdge_12.size(); ++n)
+    for (size_t n = 0; n < iEdge_12.size(); ++n)
     {
         map.push_back(iEdge_12[n]);
     }
 
-    for (unsigned int n = 0; n < iEdge_23.size(); ++n)
+    for (size_t n = 0; n < iEdge_23.size(); ++n)
     {
         map.push_back(iEdge_23[n]);
     }
 
-    for (unsigned int n = 0; n < iEdge_30.size(); ++n)
+    for (size_t n = 0; n < iEdge_30.size(); ++n)
     {
         map.push_back(iEdge_30[n]);
     }
 
-    for (unsigned int n = 0; n < iEdge_04.size(); ++n)
+    for (size_t n = 0; n < iEdge_04.size(); ++n)
     {
         map.push_back(iEdge_04[n]);
     }
 
-    for (unsigned int n = 0; n < iEdge_14.size(); ++n)
+    for (size_t n = 0; n < iEdge_14.size(); ++n)
     {
         map.push_back(iEdge_14[n]);
     }
 
-    for (unsigned int n = 0; n < iEdge_25.size(); ++n)
+    for (size_t n = 0; n < iEdge_25.size(); ++n)
     {
         map.push_back(iEdge_25[n]);
     }
 
-    for (unsigned int n = 0; n < iEdge_35.size(); ++n)
+    for (size_t n = 0; n < iEdge_35.size(); ++n)
     {
         map.push_back(iEdge_35[n]);
     }
 
-    for (unsigned int n = 0; n < iEdge_45.size(); ++n)
+    for (size_t n = 0; n < iEdge_45.size(); ++n)
     {
         map.push_back(iEdge_45[n]);
     }
 
-    for (unsigned int n = 0; n < iFace_0123.size(); ++n)
+    for (size_t n = 0; n < iFace_0123.size(); ++n)
     {
         map.push_back(iFace_0123[n]);
     }
 
-    for (unsigned int n = 0; n < iFace_014.size(); ++n)
+    for (size_t n = 0; n < iFace_014.size(); ++n)
     {
         map.push_back(iFace_014[n]);
     }
 
-    for (unsigned int n = 0; n < iFace_1254.size(); ++n)
+    for (size_t n = 0; n < iFace_1254.size(); ++n)
     {
         map.push_back(iFace_1254[n]);
     }
 
-    for (unsigned int n = 0; n < iFace_325.size(); ++n)
+    for (size_t n = 0; n < iFace_325.size(); ++n)
     {
         map.push_back(iFace_325[n]);
     }
 
-    for (unsigned int n = 0; n < iFace_0354.size(); ++n)
+    for (size_t n = 0; n < iFace_0354.size(); ++n)
     {
         map.push_back(iFace_0354[n]);
     }
 
-    for (unsigned int n = 0; n < interiorVolumePoints.size(); ++n)
+    for (size_t n = 0; n < interiorVolumePoints.size(); ++n)
     {
         map.push_back(interiorVolumePoints[n]);
     }
@@ -399,14 +396,14 @@ void NodalPrismElec::NodalPointReorder3d()
     points[1] = Array<OneD, NekDouble>(GetTotNumPoints());
     points[2] = Array<OneD, NekDouble>(GetTotNumPoints());
 
-    for (unsigned int index = 0; index < map.size(); ++index)
+    for (size_t index = 0; index < map.size(); ++index)
     {
         points[0][index] = m_points[0][index];
         points[1][index] = m_points[1][index];
         points[2][index] = m_points[2][index];
     }
 
-    for (unsigned int index = 0; index < map.size(); ++index)
+    for (size_t index = 0; index < map.size(); ++index)
     {
         m_points[0][index] = points[0][map[index]];
         m_points[1][index] = points[1][map[index]];
