+++ library/LibUtilities/Foundations/NodalTriElecData.h
@@ -36,15 +36,15 @@ namespace Nektar
 {
 namespace LibUtilities
 {
-static const unsigned int perm3A_2d[3][3] = {
+static const size_t perm3A_2d[3][3] = {
     {0, 1, 2}, {2, 0, 1}, {0, 2, 1}}; // Works for aab
-static const unsigned int perm3B_2d[3][3] = {
+static const size_t perm3B_2d[3][3] = {
     {0, 1, 2}, {1, 0, 2}, {1, 2, 0}}; // Works for abb
-static const unsigned int perm6_2d[6][3] = {
+static const size_t perm6_2d[6][3] = {
     {0, 1, 2}, {1, 0, 2}, {2, 0, 1}, // Works for abc
     {2, 1, 0}, {0, 2, 1}, {1, 2, 0}};
-const unsigned int NodalTriElecAvailable                          = 16;
-static const unsigned int NodalTriElecNPTS[NodalTriElecAvailable] = {
+const size_t NodalTriElecAvailable                          = 16;
+static const size_t NodalTriElecNPTS[NodalTriElecAvailable] = {
     1, 2, 3, 4, 5, 7, 8, 10, 12, 14, 16, 19, 21, 24, 27, 30};
 static const NekDouble NodalTriElecData[][6] = {
