+++ library/LibUtilities/Foundations/NodalTriFeketeData.h
@@ -36,17 +36,17 @@ namespace Nektar
 {
 namespace LibUtilities
 {
-static const unsigned int perm3A_2d[3][3] = {
+static const size_t perm3A_2d[3][3] = {
     {0, 1, 2}, {2, 0, 1}, {0, 2, 1}}; // Works for aab
-static const unsigned int perm3B_2d[3][3] = {
+static const size_t perm3B_2d[3][3] = {
     {0, 1, 2}, {1, 0, 2}, {1, 2, 0}}; // Works for abb
-static const unsigned int perm3C_2d[3][3] = {
+static const size_t perm3C_2d[3][3] = {
     {0, 1, 2}, {2, 0, 1}, {1, 2, 0}}; // Works for aab and abb
-static const unsigned int perm6_2d[6][3] = {
-    {0, 1, 2}, {1, 0, 2}, {2, 0, 1},
-    {2, 1, 0}, {0, 2, 1}, {1, 2, 0}}; // Works for abc
-const unsigned int NodalTriFeketeAvailable                            = 16;
-static const unsigned int NodalTriFeketeNPTS[NodalTriFeketeAvailable] = {
+static const size_t perm6_2d[6][3]   = {{0, 1, 2}, {1, 0, 2},
+                                      {2, 0, 1}, {2, 1, 0},
+                                      {0, 2, 1}, {1, 2, 0}}; // Works for abc
+const size_t NodalTriFeketeAvailable = 16;
+static const size_t NodalTriFeketeNPTS[NodalTriFeketeAvailable] = {
     1, 2, 3, 4, 5, 7, 8, 10, 12, 14, 16, 19, 21, 24, 27, 30};
 static const NekDouble NodalTriFeketeData[][6] = {
     // %%% n_1   n_3   n_6          l_1            l_2           l_3
